+++ pyneoinstance/__init__.py
@@ -1,5 +1,5 @@
 from .database.neo4jdbms import Neo4jInstance
 from .fileload.loadyaml import load_yaml_file
 
-__version__ = '4.0.2'
+__version__ = '4.0.3'
 __all__ = ['Neo4jInstance', 'load_yaml_file']